refactor(transacciones): descomponer ExpenseForm en secciones reutilizables
