+++ src/components/ExpenseForm.tsx
@@ -2,32 +2,19 @@
 
 import { useState, useEffect } from "react"
 import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select"
-import { ArrowDown, ArrowUp, CalendarIcon, CreditCard, Plus, Loader2, RefreshCw } from "lucide-react"
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
+import { Loader2 } from "lucide-react"
 import { toast } from "sonner"
 import { useCurrency } from "@/contexts/CurrencyContext"
 import { parseDDMMYYYY, formatDateToDDMMYYYY } from "@/lib/transactions/date-utils"
-import { normalizeAmountInput, formatAmountFromDigits, amountDigitsToNumber } from "@/lib/transactions/amount-utils"
+import { normalizeAmountInput, formatAmountFromDigits } from "@/lib/transactions/amount-utils"
+import { TransactionBaseFields } from "@/components/transactions/expense-form/TransactionBaseFields"
+import { RecurrentExpenseLinkSection } from "@/components/transactions/expense-form/RecurrentExpenseLinkSection"
+import { CardFinancingSection } from "@/components/transactions/expense-form/CardFinancingSection"
+import { AdvancedDateSection } from "@/components/transactions/expense-form/AdvancedDateSection"
+import { Categoria, GastoRecurrente } from "@/components/transactions/expense-form/types"
 
 interface ExpenseFormProps {
   onTransactionAdded: () => void
-}
-
-// Nueva interfaz para categorías
-interface Categoria {
-  id: number;
-  descripcion: string;
-  grupo_categoria: string | null;
-  status: boolean;
 }
 
 export function ExpenseForm({ onTransactionAdded }: ExpenseFormProps) {
@@ -53,10 +40,10 @@ export function ExpenseForm({ onTransactionAdded }: ExpenseFormProps) {
   const [tarjetaEspecifica, setTarjetaEspecifica] = useState<string>("")
 
   // NUEVO: Estados para asociar a gastos recurrentes
-  const [gastosRecurrentes, setGastosRecurrentes] = useState<any[]>([])
+  const [gastosRecurrentes, setGastosRecurrentes] = useState<GastoRecurrente[]>([])
   const [gastoRecurrenteId, setGastoRecurrenteId] = useState<string>("")
   const [loadingRecurrentes, setLoadingRecurrentes] = useState(false)
-  const [gastoRecurrenteSeleccionado, setGastoRecurrenteSeleccionado] = useState<any>(null)
+  const [gastoRecurrenteSeleccionado, setGastoRecurrenteSeleccionado] = useState<GastoRecurrente | null>(null)
 
   // Estado para controlar la visibilidad de opciones avanzadas
   const [showAdvancedOptions, setShowAdvancedOptions] = useState<boolean>(false)
@@ -312,350 +299,52 @@ export function ExpenseForm({ onTransactionAdded }: ExpenseFormProps) {
       )}
       
       <form onSubmit={handleSubmit} className="space-y-4">
-        <div className="space-y-2">
-          <Label>Tipo de Transacción</Label>
-          <RadioGroup
-            defaultValue="expense"
-            value={transactionType}
-            onValueChange={(value) => setTransactionType(value as "income" | "expense")}
-            className="flex"
-          >
-            <div className="flex items-center space-x-2 mr-4">
-              <RadioGroupItem value="expense" id="expense" />
-              <Label htmlFor="expense" className="flex items-center">
-                <ArrowDown className="mr-1 h-4 w-4 text-red-500" />
-                Gasto
-              </Label>
-            </div>
-            <div className="flex items-center space-x-2">
-              <RadioGroupItem value="income" id="income" />
-              <Label htmlFor="income" className="flex items-center">
-                <ArrowUp className="mr-1 h-4 w-4 text-green-500" />
-                Ingreso
-              </Label>
-            </div>
-          </RadioGroup>
-        </div>
+        <TransactionBaseFields
+          transactionType={transactionType}
+          setTransactionType={setTransactionType}
+          amount={amount}
+          handleAmountChange={handleAmountChange}
+          currency={currency}
+          loadingCategorias={loadingCategorias}
+          categorias={categorias}
+          movementType={movementType}
+          setMovementType={setMovementType}
+        />
 
-        <div className="space-y-2">
-          <Label htmlFor="amount">Monto</Label>
-          <div className="relative">
-            <span className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-500">
-              {currency === 'ARS' ? '$' : 'US$'}
-            </span>
-            <Input
-              id="amount"
-              name="amount"
-              type="text"
-              value={amount}
-              onChange={handleAmountChange}
-              className="pl-8"
-              placeholder="0.00"
-              required
-            />
-          </div>
-        </div>
+        <RecurrentExpenseLinkSection
+          gastoRecurrenteId={gastoRecurrenteId}
+          handleGastoRecurrenteChange={handleGastoRecurrenteChange}
+          loadingRecurrentes={loadingRecurrentes}
+          gastosRecurrentes={gastosRecurrentes}
+          gastoRecurrenteSeleccionado={gastoRecurrenteSeleccionado}
+        />
 
-        <div className="space-y-2">
-          <Label htmlFor="concepto">Concepto</Label>
-          <Input id="concepto" name="concepto" placeholder="Ej: Supermercado" required />
-        </div>
+        <CardFinancingSection
+          movementType={movementType}
+          tarjetaEspecifica={tarjetaEspecifica}
+          setTarjetaEspecifica={setTarjetaEspecifica}
+          cantidadCuotas={cantidadCuotas}
+          setCantidadCuotas={setCantidadCuotas}
+          fechaPrimerPagoStr={fechaPrimerPagoStr}
+          setFechaPrimerPagoStr={setFechaPrimerPagoStr}
+          diaPago={diaPago}
+          setDiaPago={setDiaPago}
+          amount={amount}
+          formatMoney={formatMoney}
+        />
 
-        <div className="space-y-2">
-          <Label htmlFor="categoria">Categoría</Label>
-          <Select name="categoriaId" required>
-            <SelectTrigger>
-              <SelectValue placeholder="Seleccionar categoría" />
-            </SelectTrigger>
-            <SelectContent>
-              {loadingCategorias ? (
-                <SelectItem value="loading" disabled>Cargando categorías...</SelectItem>
-              ) : categorias.length > 0 ? (
-                categorias
-                  .sort((a, b) => a.descripcion.localeCompare(b.descripcion)) // ✅ Ordenamiento alfabético
-                  .map((categoria) => (
-                    <SelectItem key={categoria.id} value={categoria.id.toString()}>
-                      {categoria.descripcion}
-                      {categoria.grupo_categoria && (
-                        <span className="text-xs text-muted-foreground ml-2">
-                          ({categoria.grupo_categoria})
-                        </span>
-                      )}
-                    </SelectItem>
-                  ))
-              ) : (
-                <>
-                  <SelectItem value="1">Alimentación</SelectItem>
-                  <SelectItem value="2">Transporte</SelectItem>
-                  <SelectItem value="3">Servicios</SelectItem>
-                  <SelectItem value="4">Ocio</SelectItem>
-                  <SelectItem value="5">Otros</SelectItem>
-                </>
-              )}
-            </SelectContent>
-          </Select>
-        </div>
-
-        {/* NUEVO: Selector de Gasto Recurrente */}
-        <div className="space-y-2">
-          <Label htmlFor="gastoRecurrente" className="flex items-center">
-            <RefreshCw className="mr-2 h-4 w-4" />
-            Asociar a Gasto Recurrente (opcional)
-          </Label>
-          <Select value={gastoRecurrenteId || "none"} onValueChange={handleGastoRecurrenteChange}>
-            <SelectTrigger>
-              <SelectValue placeholder="Seleccionar gasto recurrente..." />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="none">Ninguno</SelectItem>
-              {loadingRecurrentes ? (
-                <SelectItem value="loading" disabled>Cargando gastos recurrentes...</SelectItem>
-              ) : (
-                gastosRecurrentes.map((recurrente) => (
-                  <SelectItem key={recurrente.id} value={recurrente.id.toString()}>
-                    <div className="flex flex-col">
-                      <div className="flex items-center gap-2">
-                        <div className="font-medium">{recurrente.concepto}</div>
-                        {/* Badge de estado */}
-                        <span className={`px-2 py-0.5 text-xs rounded-full ${
-                          recurrente.estadoVisual === 'pagado' 
-                            ? 'bg-green-100 text-green-800 dark:bg-green-900/20 dark:text-green-400' 
-                            : recurrente.estadoVisual === 'pago_parcial'
-                            ? 'bg-yellow-100 text-yellow-800 dark:bg-yellow-900/20 dark:text-yellow-400'
-                            : 'bg-red-100 text-red-800 dark:bg-red-900/20 dark:text-red-400'
-                        }`}>
-                          {recurrente.estadoTexto}
-                        </span>
-                      </div>
-                      <div className="text-xs text-muted-foreground">
-                        ${recurrente.monto.toLocaleString()} total
-                        {recurrente.saldoPendiente > 0 && (
-                          <span className="ml-1">
-                            - ${Math.abs(recurrente.saldoPendiente).toLocaleString()} pendiente
-                          </span>
-                        )}
-                        {recurrente.saldoPendiente < 0 && (
-                          <span className="ml-1 text-green-600">
-                            - ${Math.abs(recurrente.saldoPendiente).toLocaleString()} sobrepagado
-                          </span>
-                        )}
-                      </div>
-                    </div>
-                  </SelectItem>
-                ))
-              )}
-            </SelectContent>
-          </Select>
-          
-          {/* Información del gasto recurrente seleccionado */}
-          {gastoRecurrenteSeleccionado && (
-            <div className="p-3 bg-blue-50 dark:bg-blue-900/20 border border-blue-200 dark:border-blue-700 rounded-md">
-              <div className="text-sm">
-                <div className="flex items-center gap-2 mb-2">
-                  <div className="font-medium text-blue-900 dark:text-blue-100">
-                    {gastoRecurrenteSeleccionado.concepto}
-                  </div>
-                  <span className={`px-2 py-0.5 text-xs rounded-full ${
-                    gastoRecurrenteSeleccionado.estadoVisual === 'pagado' 
-                      ? 'bg-green-100 text-green-800 dark:bg-green-900/20 dark:text-green-400' 
-                      : gastoRecurrenteSeleccionado.estadoVisual === 'pago_parcial'
-                      ? 'bg-yellow-100 text-yellow-800 dark:bg-yellow-900/20 dark:text-yellow-400'
-                      : 'bg-red-100 text-red-800 dark:bg-red-900/20 dark:text-red-400'
-                  }`}>
-                    {gastoRecurrenteSeleccionado.estadoTexto}
-                  </span>
-                </div>
-                <div className="text-blue-700 dark:text-blue-300 mt-1">
-                  Monto total: ${gastoRecurrenteSeleccionado.monto.toLocaleString()}
-                </div>
-                {gastoRecurrenteSeleccionado.totalPagado > 0 && (
-                  <div className="text-blue-700 dark:text-blue-300">
-                    Ya pagado: ${gastoRecurrenteSeleccionado.totalPagado.toLocaleString()} 
-                    ({gastoRecurrenteSeleccionado.porcentajePagado.toFixed(1)}%)
-                  </div>
-                )}
-                {gastoRecurrenteSeleccionado.saldoPendiente > 0 && (
-                  <div className="text-amber-700 dark:text-amber-300">
-                    Saldo pendiente: ${gastoRecurrenteSeleccionado.saldoPendiente.toLocaleString()}
-                  </div>
-                )}
-                {gastoRecurrenteSeleccionado.saldoPendiente < 0 && (
-                  <div className="text-green-700 dark:text-green-300">
-                    Sobrepagado: ${Math.abs(gastoRecurrenteSeleccionado.saldoPendiente).toLocaleString()}
-                  </div>
-                )}
-                <div className="text-xs text-blue-600 dark:text-blue-400 mt-2">
-                  💡 Este pago se asociará al gasto recurrente. 
-                  {gastoRecurrenteSeleccionado.estadoVisual === 'pagado' ? 
-                    ' Útil para facturas adicionales o imponderables.' :
-                    ' Actualizará el estado automáticamente.'
-                  }
-                </div>
-              </div>
-            </div>
-          )}
-        </div>
-
-        <div className="space-y-2">
-          <Label htmlFor="movementType">Tipo de Movimiento</Label>
-          <Select 
-            value={movementType} 
-            onValueChange={(value) => setMovementType(value as "efectivo" | "digital" | "ahorro" | "tarjeta")}
-          >
-            <SelectTrigger>
-              <SelectValue placeholder="Seleccionar tipo" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="efectivo">Efectivo</SelectItem>
-              <SelectItem value="digital">Digital</SelectItem>
-              <SelectItem value="ahorro">Ahorro</SelectItem>
-              <SelectItem value="tarjeta">
-                <div className="flex items-center">
-                  <CreditCard className="mr-1 h-4 w-4" />
-                  Tarjeta
-                </div>
-              </SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-
-        {/* Campos adicionales para financiación con tarjeta */}
-        {movementType === "tarjeta" && (
-          <div className="space-y-4 p-4 border border-gray-200 dark:border-gray-700 rounded-md bg-gray-50 dark:bg-gray-800">
-            <h4 className="font-medium text-gray-900 dark:text-gray-100">Detalles de Financiación</h4>
-            
-            <div className="space-y-2">
-              <Label htmlFor="tarjetaEspecifica">Tarjeta Específica</Label>
-              <Select value={tarjetaEspecifica} onValueChange={setTarjetaEspecifica}>
-                <SelectTrigger>  
-                  <SelectValue placeholder="Seleccionar tarjeta" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="Visa Macro">💳 Visa Macro</SelectItem>
-                  <SelectItem value="Visa Ciudad">💳 Visa Ciudad</SelectItem>
-                  <SelectItem value="Mastercard BBVA">💳 Mastercard BBVA</SelectItem>
-                  <SelectItem value="Mastercard Galicia">💳 Mastercard Galicia</SelectItem>
-                  <SelectItem value="American Express">💳 American Express</SelectItem>
-                  <SelectItem value="Naranja">🧡 Naranja</SelectItem>
-                  <SelectItem value="Cabal">💙 Cabal</SelectItem>
-                  <SelectItem value="Otra">💳 Otra tarjeta</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            
-            <div className="space-y-2">
-              <Label htmlFor="cantidadCuotas">Cantidad de Cuotas</Label>
-              <Input
-                id="cantidadCuotas"
-                type="number"
-                min="1"
-                value={cantidadCuotas}
-                onChange={(e) => setCantidadCuotas(e.target.value)}
-                placeholder="Ej: 12"
-              />
-            </div>
-            
-            <div className="space-y-2">
-              <Label htmlFor="fechaPrimerPago">Fecha del Primer Pago (DD/MM/YYYY)</Label>
-              <Input
-                type="text"
-                id="fechaPrimerPago"
-                value={fechaPrimerPagoStr}
-                onChange={(e) => setFechaPrimerPagoStr(e.target.value)}
-                placeholder="DD/MM/YYYY"
-              />
-            </div>
-            
-            <div className="space-y-2">
-              <Label htmlFor="diaPago">Día de Pago Mensual</Label>
-              <Input
-                id="diaPago"
-                type="number"
-                min="1"
-                max="31"
-                value={diaPago}
-                onChange={(e) => setDiaPago(e.target.value)}
-                placeholder="Ej: 10"
-              />
-              <p className="text-xs text-gray-500 dark:text-gray-400">
-                Día del mes en que se realiza el pago (opcional)
-              </p>
-            </div>
-            
-            <div className="pt-2">
-              <p className="text-sm text-gray-600 dark:text-gray-300">
-                Monto por cuota: {cantidadCuotas && amount ? 
-                  formatMoney(amountDigitsToNumber(amount) / parseInt(cantidadCuotas)) : 
-                  formatMoney(0)}
-              </p>
-            </div>
-          </div>
-        )}
-
-        {/* Botón para mostrar/ocultar opciones avanzadas */}
-        <Button 
-          type="button" 
-          variant="outline" 
-          className="w-full flex items-center justify-center"
-          onClick={() => setShowAdvancedOptions(!showAdvancedOptions)}
-        >
-          {showAdvancedOptions ? "Ocultar opciones avanzadas" : "Mostrar opciones avanzadas"}
-          <Plus className={`ml-2 h-4 w-4 transition-transform ${showAdvancedOptions ? "rotate-45" : ""}`} />
-        </Button>
-
-        {/* Contenido visible siempre */}
-        <div className={`space-y-4 ${showAdvancedOptions ? "" : "hidden"}`}>
-          <div className="space-y-2">
-            <Label htmlFor="date">Fecha de depósito/transacción (DD/MM/YYYY)</Label>
-            <Input
-              type="text"
-              id="date"
-              value={dateStr}
-              onChange={(e) => {
-                setDateStr(e.target.value)
-                const parsedDate = parseDDMMYYYY(e.target.value)
-                setDate(parsedDate)
-              }}
-              placeholder="DD/MM/YYYY"
-            />
-          </div>
-
-          {/* Checkbox para activar fecha de imputación */}
-          <div className="flex items-center space-x-2">
-            <input
-              type="checkbox"
-              id="mostrarFechaImputacion"
-              checked={mostrarFechaImputacion}
-              onChange={(e) => setMostrarFechaImputacion(e.target.checked)}
-              className="h-4 w-4 text-blue-600 focus:ring-blue-500 border-gray-300 rounded"
-            />
-            <Label htmlFor="mostrarFechaImputacion" className="flex items-center cursor-pointer">
-              <CalendarIcon className="mr-2 h-4 w-4" />
-              Usar fecha diferente para imputación contable
-            </Label>
-          </div>
-
-          {/* Campo de fecha de imputación */}
-          {mostrarFechaImputacion && (
-            <div className="space-y-2 p-4 border border-amber-200 dark:border-amber-700 rounded-md bg-amber-50 dark:bg-amber-900/20">
-              <Label htmlFor="fechaImputacion">Fecha de imputación contable (DD/MM/YYYY)</Label>
-              <Input
-                type="text"
-                id="fechaImputacion"
-                value={fechaImputacionStr}
-                onChange={(e) => {
-                  setFechaImputacionStr(e.target.value)
-                  const parsedFechaImputacion = parseDDMMYYYY(e.target.value)
-                  setFechaImputacion(parsedFechaImputacion)
-                }}
-                placeholder="DD/MM/YYYY"
-              />
-              <p className="text-xs text-amber-700 dark:text-amber-300">
-                <strong>Ejemplo:</strong> Salario depositado el 31/05/2024 pero corresponde a junio → usar 01/06/2024 como fecha de imputación
-              </p>
-            </div>
-          )}
-        </div>
+        <AdvancedDateSection
+          showAdvancedOptions={showAdvancedOptions}
+          setShowAdvancedOptions={setShowAdvancedOptions}
+          dateStr={dateStr}
+          setDateStr={setDateStr}
+          setDate={setDate}
+          mostrarFechaImputacion={mostrarFechaImputacion}
+          setMostrarFechaImputacion={setMostrarFechaImputacion}
+          fechaImputacionStr={fechaImputacionStr}
+          setFechaImputacionStr={setFechaImputacionStr}
+          setFechaImputacion={setFechaImputacion}
+        />
 
         <Button type="submit" className="w-full" disabled={loading}>
           {loading ? (
